backend-common: Remove read method from UrlReader
