+++ packages/backend-common/api-report.md
@@ -33,9 +33,19 @@ import { MergeResult } from 'isomorphic-git';
 import { PushResult } from 'isomorphic-git';
 import { Readable } from 'stream';
 import { ReadCommitResult } from 'isomorphic-git';
+import { ReadTreeOptions } from '@backstage/backend-plugin-api';
+import { ReadTreeResponse } from '@backstage/backend-plugin-api';
+import { ReadTreeResponseDirOptions } from '@backstage/backend-plugin-api';
+import { ReadTreeResponseFile } from '@backstage/backend-plugin-api';
+import { ReadUrlOptions } from '@backstage/backend-plugin-api';
+import { ReadUrlResponse } from '@backstage/backend-plugin-api';
 import { RequestHandler } from 'express';
 import { Router } from 'express';
+import { SearchOptions } from '@backstage/backend-plugin-api';
+import { SearchResponse } from '@backstage/backend-plugin-api';
+import { SearchResponseFile } from '@backstage/backend-plugin-api';
 import { Server } from 'http';
+import { UrlReaderService as UrlReader } from '@backstage/backend-plugin-api';
 import { V1PodTemplateSpec } from '@kubernetes/client-node';
 import * as winston from 'winston';
 import { Writable } from 'stream';
@@ -533,30 +543,11 @@ export type ReaderFactory = (options: {
   treeResponseFactory: ReadTreeResponseFactory;
 }) => UrlReaderPredicateTuple[];
 
-// @public
-export type ReadTreeOptions = {
-  filter?(
-    path: string,
-    info?: {
-      size: number;
-    },
-  ): boolean;
-  etag?: string;
-  signal?: AbortSignal;
-};
+export { ReadTreeOptions };
 
-// @public
-export type ReadTreeResponse = {
-  files(): Promise<ReadTreeResponseFile[]>;
-  archive(): Promise<NodeJS.ReadableStream>;
-  dir(options?: ReadTreeResponseDirOptions): Promise<string>;
-  etag: string;
-};
+export { ReadTreeResponse };
 
-// @public
-export type ReadTreeResponseDirOptions = {
-  targetDir?: string;
-};
+export { ReadTreeResponseDirOptions };
 
 // @public
 export interface ReadTreeResponseFactory {
@@ -587,24 +578,11 @@ export type ReadTreeResponseFactoryOptions = {
   ) => boolean;
 };
 
-// @public
-export type ReadTreeResponseFile = {
-  path: string;
-  content(): Promise<Buffer>;
-};
+export { ReadTreeResponseFile };
 
-// @public
-export type ReadUrlOptions = {
-  etag?: string;
-  signal?: AbortSignal;
-};
+export { ReadUrlOptions };
 
-// @public
-export type ReadUrlResponse = {
-  buffer(): Promise<Buffer>;
-  stream?(): Readable;
-  etag?: string;
-};
+export { ReadUrlResponse };
 
 // @public
 export class ReadUrlResponseFactory {
@@ -652,23 +630,11 @@ export type RunContainerOptions = {
   pullImage?: boolean;
 };
 
-// @public
-export type SearchOptions = {
-  etag?: string;
-  signal?: AbortSignal;
-};
+export { SearchOptions };
 
-// @public
-export type SearchResponse = {
-  files: SearchResponseFile[];
-  etag: string;
-};
+export { SearchResponse };
 
-// @public
-export type SearchResponseFile = {
-  url: string;
-  content(): Promise<Buffer>;
-};
+export { SearchResponseFile };
 
 // @public
 export class ServerTokenManager implements TokenManager {
@@ -755,13 +721,7 @@ export interface TokenManager {
   }>;
 }
 
-// @public
-export type UrlReader = {
-  read(url: string): Promise<Buffer>;
-  readUrl(url: string, options?: ReadUrlOptions): Promise<ReadUrlResponse>;
-  readTree(url: string, options?: ReadTreeOptions): Promise<ReadTreeResponse>;
-  search(url: string, options?: SearchOptions): Promise<SearchResponse>;
-};
+export { UrlReader };
 
 // @public
 export type UrlReaderPredicateTuple = {
