+++ packages/backend-common/src/reading/BitbucketServerUrlReader.ts
@@ -41,7 +41,7 @@ import {
 import { ReadUrlResponseFactory } from './ReadUrlResponseFactory';
 
 /**
- * Implements a {@link UrlReader} for files from Bitbucket Server APIs.
+ * Implements a {@link @backstage/backend-plugin-api#UrlReaderService} for files from Bitbucket Server APIs.
  *
  * @public
  */
