+++ packages/backend-common/src/reading/GiteaUrlReader.ts
@@ -36,7 +36,7 @@ import {
 import { Readable } from 'stream';
 
 /**
- * Implements a {@link UrlReader} for the Gitea v1 api.
+ * Implements a {@link @backstage/backend-plugin-api#UrlReaderService} for the Gitea v1 api.
  *
  * @public
  */
