+++ packages/backend-plugin-api/api-report.md
@@ -3,6 +3,8 @@
 > Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
 
 ```ts
+/// <reference types="node" />
+
 import { Config } from '@backstage/config';
 import { Handler } from 'express';
 import { Logger } from 'winston';
@@ -12,9 +14,9 @@ import { PluginCacheManager } from '@backstage/backend-common';
 import { PluginDatabaseManager } from '@backstage/backend-common';
 import { PluginEndpointDiscovery } from '@backstage/backend-common';
 import { PluginTaskScheduler } from '@backstage/backend-tasks';
+import { Readable } from 'stream';
 import { TokenManager } from '@backstage/backend-common';
 import { TransportStreamOptions } from 'winston-transport';
-import { UrlReader } from '@backstage/backend-common';
 
 // @public (undocumented)
 export interface BackendFeature {
@@ -245,6 +247,50 @@ export interface PluginMetadataService {
 // @public (undocumented)
 const pluginMetadataServiceRef: ServiceRef<PluginMetadataService, 'plugin'>;
 
+// @public
+export type ReadTreeOptions = {
+  filter?(
+    path: string,
+    info?: {
+      size: number;
+    },
+  ): boolean;
+  etag?: string;
+  signal?: AbortSignal;
+};
+
+// @public
+export type ReadTreeResponse = {
+  files(): Promise<ReadTreeResponseFile[]>;
+  archive(): Promise<NodeJS.ReadableStream>;
+  dir(options?: ReadTreeResponseDirOptions): Promise<string>;
+  etag: string;
+};
+
+// @public
+export type ReadTreeResponseDirOptions = {
+  targetDir?: string;
+};
+
+// @public
+export type ReadTreeResponseFile = {
+  path: string;
+  content(): Promise<Buffer>;
+};
+
+// @public
+export type ReadUrlOptions = {
+  etag?: string;
+  signal?: AbortSignal;
+};
+
+// @public
+export type ReadUrlResponse = {
+  buffer(): Promise<Buffer>;
+  stream?(): Readable;
+  etag?: string;
+};
+
 // @public (undocumented)
 export type RootLoggerService = LoggerService;
 
@@ -256,6 +302,24 @@ export type SchedulerService = PluginTaskScheduler;
 
 // @public (undocumented)
 const schedulerServiceRef: ServiceRef<PluginTaskScheduler, 'plugin'>;
+
+// @public
+export type SearchOptions = {
+  etag?: string;
+  signal?: AbortSignal;
+};
+
+// @public
+export type SearchResponse = {
+  files: SearchResponseFile[];
+  etag: string;
+};
+
+// @public
+export type SearchResponseFile = {
+  url: string;
+  content(): Promise<Buffer>;
+};
 
 // @public (undocumented)
 export type ServiceFactory<TService = unknown> =
@@ -307,9 +371,13 @@ export type TypesToServiceRef<T> = {
   [key in keyof T]: ServiceRef<T[key]>;
 };
 
-// @public (undocumented)
-export type UrlReaderService = UrlReader;
+// @public
+export type UrlReaderService = {
+  readUrl(url: string, options?: ReadUrlOptions): Promise<ReadUrlResponse>;
+  readTree(url: string, options?: ReadTreeOptions): Promise<ReadTreeResponse>;
+  search(url: string, options?: SearchOptions): Promise<SearchResponse>;
+};
 
 // @public (undocumented)
-const urlReaderServiceRef: ServiceRef<UrlReader, 'plugin'>;
+const urlReaderServiceRef: ServiceRef<UrlReaderService, 'plugin'>;
 ```
