+++ plugins/azure-devops-backend/src/utils/azure-devops-utils.test.ts
@@ -255,7 +255,6 @@ describe('replaceReadme', () => {
         }),
       readTree: jest.fn(),
       search: jest.fn(),
-      read: jest.fn(),
     };
 
     const result = await replaceReadme(
