+++ plugins/scaffolder-backend-module-cookiecutter/src/actions/fetch/cookiecutter.test.ts
@@ -75,7 +75,6 @@ describe('fetch:cookiecutter', () => {
 
   const mockReader: UrlReader = {
     readUrl: jest.fn(),
-    read: jest.fn(),
     readTree: jest.fn(),
     search: jest.fn(),
   };
