+++ plugins/scaffolder-backend-module-rails/src/actions/fetch/rails/index.test.ts
@@ -74,7 +74,6 @@ describe('fetch:rails', () => {
   };
 
   const mockReader: UrlReader = {
-    read: jest.fn(),
     readUrl: jest.fn(),
     readTree: jest.fn(),
     search: jest.fn(),
