+++ plugins/scaffolder-backend/src/scaffolder/actions/builtin/fetch/plain.test.ts
@@ -34,7 +34,6 @@ describe('fetch:plain', () => {
   );
   const reader: UrlReader = {
     readUrl: jest.fn(),
-    read: jest.fn(),
     readTree: jest.fn(),
     search: jest.fn(),
   };
