+++ plugins/techdocs-backend/src/service/standaloneServer.ts
@@ -53,7 +53,6 @@ export async function startStandaloneServer(
   });
   const discovery = SingleHostDiscovery.fromConfig(config);
   const mockUrlReader: jest.Mocked<UrlReader> = {
-    read: jest.fn(),
     readUrl: jest.fn(),
     readTree: jest.fn(),
     search: jest.fn(),
