+++ plugins/techdocs-node/src/stages/prepare/dir.test.ts
@@ -45,7 +45,6 @@ const createMockEntity = (annotations: {}) => {
 
 const mockConfig = new ConfigReader({});
 const mockUrlReader: jest.Mocked<UrlReader> = {
-  read: jest.fn(),
   readUrl: jest.fn(),
   readTree: jest.fn(),
   search: jest.fn(),
